+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- While in `telescope` buffers popup
+  - In `normal` mode, delete buffer using `d`
+  - In `insert` mode, delete buffer using `Ctrl+D`
+
 ## [v0.4.0]
 
 ### Added